requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
Django @ file:///tmp/build/80754af9/django_1600355803933/work
djangorestframework==3.11.1
Pillow==7.2.0
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
